+++ packages/app/src/pages/Pools/ClosurePrompts.tsx
@@ -67,7 +67,7 @@ export const ClosurePrompts = () => {
                 }
                 onClick={() =>
                   openModal({
-                    key: 'Unbond',
+                    key: 'LeavePool',
                     options: { bondFor: 'pool' },
                     size: 'sm',
                   })
